Update: move the nms package to util, and fix a bug
