--- common/src/main/java/cc/zoyn/core/util/nms/package-info.java
@@ -1,7 +1,0 @@
-/**
- * The NMS of utils
- *
- * @author Zoyn
- * @since 2017-12-09
- */
-package cc.zoyn.core.util.nms;
